Refactor PathTracker component and introduce new subcomponents for improved structure and readability. Added StationSelector, StatusRibbon, and ArrivalsTable components. Implemented custom hook usePathData for data fetching and state management. Enhanced code organization by moving constants and utility functions to dedicated files.
