+++ src/services/pathApi.ts
@@ -1,0 +1,16 @@
+import { RidePathResponse } from "../types/path";
+import { RAW_API_URL, PROXY_API_URL } from "../constants/stations";
+
+export const fetchRidePath = async (): Promise<RidePathResponse> => {
+  const attempt = async (url: string) => {
+    const res = await fetch(url, { cache: "no-store" });
+    if (!res.ok) throw new Error(`HTTP ${res.status}`);
+    return res.json();
+  };
+
+  try {
+    return await attempt(RAW_API_URL);
+  } catch {
+    return await attempt(PROXY_API_URL);
+  }
+};
